--- a/Project.toml
+++ b/Project.toml
@@ -14,13 +14,13 @@ StaticArrays = "90137ffa-7385-5640-81b9-e52037218182"
 WriteVTK = "64499a7a-5c06-52f2-abe2-ccb03c286192"
 
 [compat]
-julia = "1.7"
 AbaqusReader = "0.2"
 FileIO = "1"
 JLD2 = "0.4"
 ProgressMeter = "1"
 StaticArrays = "1"
 WriteVTK = "1"
+julia = "1.7"
 
 [extras]
 SafeTestsets = "1bc83da4-3b8d-516f-aca4-4fe02f6d838f"