--- a/test/contact.jl
+++ b/test/contact.jl
@@ -36,18 +36,18 @@ if Threads.nthreads() <= 2
     rm.(joinpath.(@__DIR__,filter(x->endswith(x,".vtu"), readdir(@__DIR__))), force=true)
     rm.(joinpath.(@__DIR__,filter(x->endswith(x,".jld2"), readdir(@__DIR__))), force=true)
     rm.(joinpath.(@__DIR__,filter(x->endswith(x,".log"), readdir(@__DIR__))), force=true)
-    resulting_bodies = submit(job1)
+    bodies = submit(job1)
 
     @test contact1.spring_constant == 1e12
     b_int_contact = 0.5 * 9 * 1e12 / (π * δ^4) * (1-0.5) / 0.5
-    @test resulting_bodies[1].b_int[1,2,1] == -b_int_contact
-    @test resulting_bodies[2].b_int[1,1,1] == b_int_contact
+    @test bodies[1].b_int[1,2,1] == -b_int_contact
+    @test bodies[2].b_int[1,1,1] == b_int_contact
     acc_contact = b_int_contact / rho
-    @test resulting_bodies[1].acceleration[1,2,1] == -acc_contact
-    @test resulting_bodies[2].acceleration[1,1,1] == acc_contact
+    @test bodies[1].acceleration[1,2,1] == -acc_contact
+    @test bodies[2].acceleration[1,1,1] == acc_contact
     vel_contact = acc_contact * job1.td.Δt * 0.5
-    @test resulting_bodies[1].velocity[1,2,1] == -vel_contact
-    @test resulting_bodies[2].velocity[1,1,1] == vel_contact
+    @test bodies[1].velocity[1,2,1] == -vel_contact
+    @test bodies[2].velocity[1,1,1] == vel_contact
     @test length(filter(x->endswith(x,".vtu"), readdir(@__DIR__))) == 4
     @test length(filter(x->endswith(x,".jld2"), readdir(@__DIR__))) == 4
     @test length(filter(x->endswith(x,".log"), readdir(@__DIR__))) == 1