The multithreading performance could be slightly improved by optimising the summation of the force densities calculated in the individual threads.

### Performance comparison:
nthreads | time old [s] | time new [s] | gain
---|---|---|---
1  | 284.61 | 293.12 | 0.97
2  | 153.14 | 150.98 | 1.01
8  |  49.91 |  46.61 | 1.07
12 |  43.71 |  38.6  | 1.13
16 |  41.72 |  35.93 | 1.16
24 |  57.4  |  37.95 | 1.51
32 |  98.26 |  49.7  | 1.98